requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aiounittest==1.4.2
amqp==5.2.0
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
billiard==4.2.0
bitarray==2.8.2
bitstring==4.1.2
blinker==1.6.2
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.2
eventlet==0.33.3
exceptiongroup==1.1.3
filelock==3.12.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gevent==23.9.1
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.16.4
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
kombu==5.3.4
lightning-utilities==0.9.0
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.25.2
opencv-python==4.8.0.76
outcome==1.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
priority==2.0.0
prompt-toolkit==3.0.41
psutil==5.9.5
py-cpuinfo==9.0.0
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.19.0
pylivestream==2.0.0
pyparsing==3.0.9
pyrtmp==0.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-engineio==4.6.0
python-socketio==5.8.0
pytorch-lightning==1.9.5
pytz==2023.3
PyYAML==6.0.1
Quart==0.19.3
redis==5.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
taskgroup==0.0.0a4
timm==0.9.5
tomli==2.0.1
torch==1.13.1+cpu
torchmetrics==1.0.3
torchvision==0.14.1+cpu
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.159
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.10
websocket-client==1.6.4
Werkzeug==3.0.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.1
